refactor: Unify Wallet/TestWallet into bdk_tx_testenv workspace crate
- Convert repo to a Cargo workspace with a new unpublished
  bdk_tx_testenv crate under testenv/
- Merge the duplicated Wallet (examples/common.rs) and TestWallet
  (tests/timelock.rs) into a single Wallet struct with signer and secp
  fields always present
- Deduplicate old_rpc_client() helper (was copied in 4 files)
- Add all_candidates_with(&Assets) method for custom asset parameters
- Reduce examples/common.rs to a thin re-export shim
