+++ Cargo.toml
@@ -1,3 +1,6 @@
+[workspace]
+members = [".", "testenv"]
+
 [package]
 name = "bdk_tx"
 version = "0.1.0"
@@ -19,6 +22,7 @@ rand = { version = "0.8", optional = true }
 [dev-dependencies]
 anyhow = "1"
 bdk_tx = { path = "." }
+bdk_tx_testenv = { path = "testenv" }
 bitcoin = { version = "0.32", default-features = false, features = ["rand-std"] }
 bdk_testenv = "0.13.0"
 bdk_bitcoind_rpc = "0.22.0"
